Run the varchar collation tests in invariant globalization mode
The invariant leg of the matrix failed with exit code 8 ('zero tests
ran'). Every test in the assembly sits in a class marked NotInvariant, so
the whole run was skipped, and CI only selects the projects a change
affects: this PR made TdsServer.Tests the sole project of its shard, and
the run had nothing left to report. MinimumExpectedTests was set to 0 for
that reason, but it only drops the argument setting the expected count;
the platform still ends the run with exit code 8, and the dotnet test
orchestrator computes that code from the aggregated results, so no
per-runner argument can suppress it.

The collation tests do not need a real culture: they drive the protocol
over a socket rather than through SqlClient. They move to their own class
that is not skipped, so the assembly always has tests to run, and the
sort id and UTF-8 paths gain the coverage they most need - those are the
collations that have to keep working when no culture data is available.
The Windows collation test keeps NotInvariant, and a counterpart asserts
the rejection that the locale round-trip guard produces in invariant
mode.

The raw client helpers move to RawTdsClient so both classes can use them,
one type per file as MA0048 requires. MinimumExpectedTests goes back to
its default, which now guards against the assembly regressing to zero
tests instead of allowing it.
